requirements.txt
dj-database-url
Django
djangorestframework
django-cors-headers
Pillow
django-allauth
django-rest-auth
factory-boy
psycopg2
whitenoise==5.0.1
gunicorn==20.0.4
django-admin-honeypot==1.1.0